requirements.txt
flask
pandas
gunicorn
google-genai